Corrected Guid also for quoted identifier quoter
Made non quoted identifier oracle quoter inherit from oracle quoted identifier quoter so we have a
clear place to split formatting methods from quote related issues.
This could indicate that the quoter should be split into two separate classes for instance a value quoter and an identifier quoter.
